examples/research/requirements.txt
deepagents
langgraph-cli[inmem]
tavily-python
